requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
mistralai==0.1.8
python-dotenv==1.0.1
python-multipart==0.0.12
requests==2.32.3
pillow==10.4.0
loguru==0.7.2
apscheduler==3.10.4